#161 refactor: recruitment post schema 분리
